// server/src/types/tournament.types.ts
export enum RoomStage {
  WAITING = "waiting",
  PERFORMANCE = "performance",
  VOTING = "voting",
  WAITING_FOR_ROUND_COMPLETION = "waiting_for_round_completion",
  RESULTS = "results",
  COMPLETED = "completed",
}

export enum ParticipantStatus {
  REGISTERED = "registered",
  ACTIVE = "active",
  PERFORMING = "performing",
  VOTING = "voting",
  WAITING_FOR_ROUND_COMPLETION = "waiting_for_round_completion",
  ELIMINATED = "eliminated",
  WINNER = "winner",
  OFFLINE = "offline",
}